Move methods to other class
