+++ src/main/java/handlers/CameraHandler.java
@@ -40,28 +40,8 @@ public class CameraHandler {
      * @param camera The camera to get the new frame from.
      * @return The new frame as a BufferedImage.
      */
-    public BufferedImage getNewFrame(Camera camera) {
-        Mat mat = camera.getLastFrame();
-        return matToBufferedImage(mat);
-    }
-
-    /**
-     * Converts a Mat to a BufferedImage.
-     * @param videoMatImage The frame in Mat.
-     * @return The BufferedImage.
-     */
-    private BufferedImage matToBufferedImage(Mat videoMatImage) {
-
-        int type = videoMatImage.channels() == 1
-            ? BufferedImage.TYPE_BYTE_GRAY : BufferedImage.TYPE_3BYTE_BGR;
-        int bufferSize = videoMatImage.channels() * videoMatImage.cols() * videoMatImage.rows();
-        byte[] buffer = new byte[bufferSize];
-        videoMatImage.get(0, 0, buffer);
-        BufferedImage image = new BufferedImage(videoMatImage.cols(), videoMatImage.rows(), type);
-        final byte[] targetPixels = ((DataBufferByte) image.getRaster().getDataBuffer()).getData();
-        System.arraycopy(buffer, 0, targetPixels, 0, buffer.length);
-        return image;
-
+    public Mat getNewFrame(Camera camera) {
+        return camera.getLastFrame();
     }
 
     public List<Camera> getCameraList() {
